fix(image-pipeline): T12 defect #2 — never give a colour channel an 8-bit type
ROOT CAUSE. On Adreno/Vulkan the split Stage-4 yuv420 kernel lost the carry out
of bit 7, conditionally, whenever two or more colour channels were accumulated:
the pre-shift sum was wrong by an exact multiple of 256 and by nothing else
(alphabet {0,±64,±128,±192} at >>2 over 155M measured pixels), on ~1/3 of the
pixels whose sum crosses a byte boundary. One gathered channel was byte-perfect
through the identical uint8 store. Every arithmetic instruction in the emitted
SPIR-V was ALREADY 32-bit (dbg-d4-08), so the defect is not expressible as an
arithmetic error in the module the driver receives; the operands were merely
BORN `OpConvertFToU %uchar` and widened back with `OpUConvert` before the adds,
and the count of those widenings (0 / 6 / 9) tracked the defect exactly across
every kernel built. Five rounds of probes varied arithmetic domain, association,
coefficients, magnitude, provable range, store form, operand count and 26 gather
permutations — all downstream of that birth, and not one moved a single pixel.

FIX. Quantise to the same integer values but never give them an 8-bit TYPE:
`encode8`'s emit (the shipping path at diag_stage = -1) and `linear8` (the diag
ladder) now `cast<int32_t>`, and the rgba8 kernel re-narrows with `cast<uint8_t>`
at its own store, which is the only narrow conversion left on the path and is
never an arithmetic operand. Value-identity is by construction, not by tolerance:
both quantisers clamp in float to [0,255] before the cast, so OpConvertFToU and
OpConvertFToS truncate to the same integer for every input. One shared body, both
arms, both backends, no platform guard, no AOT target change, no schedule change,
and zero memory delta — rgb8_r/g/b are inline Funcs, so a type change allocates
nothing.

EVIDENCE (adb cc5bf709, /data/local/tmp/archa; artifacts native/tests/tmp/archa-*):
  content, before any device run: kernel-0 store-reaching slice
    OpConvertFToU->%uchar 18 -> 0, OpUConvert-from-%uchar 18 -> 0 (archa-15);
    chroma kernels 36 -> 0 and 48 -> 12, the 12 survivors being a char-narrow
    inside an int32 table-index computation, not a colour birth.
  device luma: gross 8525983 -> 0 (xtrans-raf) and 7095334 -> 0 (linearrgb-x3f),
    Y-plane FNV now equal to the host oracle's, every difference bucket zero.
  device whole-frame: three-plane Y4b diff_bytes == 0 on both decisive routes,
    so CHROMA is byte-exact too (archa-50); the chroma gross populations on
    record were measured on the defective build and are now zero.
    bayer-arw's diff_bytes=75 @1241174 is the fused-vs-two-stage CONTROL residue
    (raw_gpu_pipeline.cpp:531-532 sends the rgba8 control through the fused
    kernel), reproduced count-for-count and offset-for-offset by an independent
    build; that attribution is the leading explanation and is NOT directly
    confirmed (archc-28 names the one build that would confirm it).
  Metal: host Y-suite 30 pass / 0 fail / 2 n/a on both arms, and the two suite
    outputs are BYTE-IDENTICAL including every FNV line — the no-op is measured.
  throughput: 0.988x vs the shipping library over interleaved reps, i.e. inside
    the shipping arm's own spread (archa-52). The change deletes conversions and
    adds none.

NOTE FOR THE NEXT READER. `linear8` is INERT in the shipping configuration — it
serves diag_stage 0..7 only, and diag_stage = -1 falls through to `encode8`.
Editing it alone produces a byte-identical AOT archive, which is indistinguishable
from "the simplifier refolded my type" unless a liveness control is run
(perturbing the clamp constant must move the module: at linear8 it does not, at
encode8 it does). It is changed here for type uniformity of out8_r/g/b, not
because it was on the defect path.

DELIBERATELY NOT INCLUDED: the same treatment at the non-split body's own
quantisation site (dng_render_stage4_expr.h:473-475) plus re-narrowing at its two
rgba8 consumers, which would extend the "no 8-bit-typed value upstream of
arithmetic" rule tree-wide. That variant was never built or measured, and no
claim is made about it; see native/tests/tmp/archa-41-VERDICT.txt section 7.

+++ native/generators/dng_render_stage4_split_expr.h
@@ -153,8 +153,18 @@ void build_dng_render_stage4_split_rgb8(
         Expr s_g = cast<float>(src_rgb(clamp(base + 1, 0, max_idx))) * src_scale;
         Expr s_b = cast<float>(src_rgb(clamp(base + 2, 0, max_idx))) * src_scale;
 
+        // ARCH2-A (validation round, Task #14). The channel is quantised to
+        // the SAME integer value but is BORN int32, never uint8. The argument
+        // is already clamped to [0.0f, 255.0f], so it is non-negative and in
+        // range for both conversions, and cast<uint8_t> / cast<int32_t>
+        // (OpConvertFToU %uchar / OpConvertFToS %int) return the same integer
+        // for every input: value-identical by construction. Consumers that
+        // need 8-bit BYTES re-narrow at their own store, which is then the
+        // only narrow type on the path and is never an arithmetic operand.
+        // (This lambda serves the diag_stage 0..7 ladder only; the shipping
+        // path's birth is the encode8 emit at the end of this function.)
         auto linear8 = [&](Expr v) {
-            return cast<uint8_t>(clamp(v * 255.0f + 0.5f, 0.0f, 255.0f));
+            return cast<int32_t>(clamp(v * 255.0f + 0.5f, 0.0f, 255.0f));
         };
 
         // Stash the three channel results; finalize() wires them into dst_rgb.
@@ -418,9 +428,12 @@ void build_dng_render_stage4_split_rgb8(
             return clamp(g * 255.0f + 0.5f, 0.0f, 255.0f);
         };
 
-        emit_rgb8(cast<uint8_t>(encode8(f_r)),
-                  cast<uint8_t>(encode8(f_g)),
-                  cast<uint8_t>(encode8(f_b)));
+        // ARCH2-A: THE SHIPPING PATH'S 8-BIT BIRTH (diag_stage = -1 ends here).
+        // encode8 already clamps to [0.0f, 255.0f], so cast<int32_t> yields
+        // the identical integer cast<uint8_t> yielded, for every input.
+        emit_rgb8(cast<int32_t>(encode8(f_r)),
+                  cast<int32_t>(encode8(f_g)),
+                  cast<int32_t>(encode8(f_b)));
 // ---- END verbatim slice -----------------------------------------------------
 }
